+++ sdl3/sunny_land/sunny_land/game/scene/game_scene.h
@@ -14,7 +14,7 @@ public:
     // 覆盖场景基类的核心方法
     void init() override;
     void handleInput() override;
-    void update(std::chrono::duration<double> delta_time) override;
+    void update(std::chrono::duration<float> delta_time) override;
     void render() override;
     void clean() override;
